--- src/helpers/compute-easy.spec.js
@@ -1,3 +1,0 @@
-test('adds 1 + 2 to equal 3', () => {
-  expect(1).toBe(1);
-});
